src/ru/job4j/ru/job4j/loop/Counter.java
package ru.job4j.ru.job4j.loop;

public class Counter {
    public static int sum(int start, int finish) {
        int sum = 0;
        int count = finish - start;
        for (int index = start; index <= finish; index++) {
            sum = sum + start + index;
        }
        return sum;
    }

    public static int sumByEven(int start, int finish) {
        int sum = 0;
        for (int index = 0; index <= finish - start; index++) {
            if ((index + start) % 2 == 0) {
                sum = sum + start + index;
            }
        }
        return sum;
    }

    public static void main(String[] args) {
        System.out.println(sum(0, 10));
        System.out.println(sum(3, 8));
        System.out.println(sum(1, 1));

        System.out.println(sumByEven(0, 10));
        System.out.println(sumByEven(3, 8));
        System.out.println(sumByEven(1, 1));
    }
}
